+++ servant-swagger-ui-core/Changelog.md
@@ -1,3 +1,7 @@
+# 0.4.0
+
+- Switch to OpenApi 3.0
+
 # 0.3.3
 
 - Add `swaggerSchemaUIServerImpl'`